recipehub_fullstack/backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.5.1,<5.6
django-cors-headers>=4.7,<5
Pillow>=10.0
